+++ heyoka/docstrings.cpp
@@ -860,6 +860,11 @@ std::string sgp4_propagator(const std::string &p)
    A :ref:`tutorial <tut_sgp4_propagator>` explaining the use of this class
    is available.
 
+.. note::
+
+   Consider using the factory function :py:func:`~heyoka.model.sgp4_propagator()`, rather
+   that constructing instances of this class by hand.
+
 )",
                        p);
 }
@@ -873,11 +878,6 @@ Constructor.
 .. versionadded:: 7.0.0
 
    This function now also accepts *sat_list* as a NumPy array.
-
-.. note::
-
-   As an alternative to this constructor, consider using the factory function
-   :py:func:`~heyoka.model.sgp4_propagator()`.
 
 The constructor will initialise the propagator from *sat_list*, which must be either a list
 of general perturbations element sets (GPEs) represented as ``Satrec`` objects from the
@@ -1211,6 +1211,10 @@ std::string eop_data()
 
 This class is used to manage and access Earth orientation parameters data.
 
+.. note::
+
+   A :ref:`tutorial <tut_eop_data>` illustrating the use of this class is available.
+
 )";
 }
 
@@ -1361,6 +1365,70 @@ axes of the FK5 frame at J2000.
 :returns: the rotated Cartesian coordinates.
 
 )";
+}
+
+std::string rot_itrs_icrs(double thresh)
+{
+    return fmt::format(
+        R"(rot_itrs_icrs(xyz: typing.Iterable[expression], time_expr: expression = heyoka.time, thresh: float = {}, eop_data: eop_data = eop_data()) -> list[expression]
+
+Rotation from ITRS to ICRS.
+
+.. versionadded:: 7.3.0
+
+This function will rotate the input Cartesian coordinates *xyz* from the axes of the
+`ITRS <https://en.wikipedia.org/wiki/International_Terrestrial_Reference_System_and_Frame>`__ frame to the axes of the
+`ICRS <https://en.wikipedia.org/wiki/International_Celestial_Reference_System_and_its_realizations>`__ frame.
+
+While the orientation of the ICRS is fixed in inertial space, the orientation of the ITRS is fixed relative to the Earth,
+and thus the rotation depends on the input time *time_expr* (which is expected to represent the number of Julian centuries elapsed
+since the epoch of J2000 in the `terrestrial time scale (TT) <https://en.wikipedia.org/wiki/Terrestrial_Time>`__).
+
+The rotation also requires in input an Earth orientation parameters dataset *eop_data* and a truncation threshold *thresh*
+for the IAU2000/2006 precession-nutation theory. Please see the documentation of :py:class:`~heyoka.eop_data` and
+:py:func:`~heyoka.model.iau2006()` for more information.
+
+:param xyz: the input Cartesian coordinates.
+:param time_expr: the input time expression.
+:param thresh: the truncation threshold for the IAU2000/2006 theory.
+:param eop_data: the EOP data to be used for the computation.
+
+:returns: the rotated Cartesian coordinates.
+
+)",
+        thresh);
+}
+
+std::string rot_icrs_itrs(double thresh)
+{
+    return fmt::format(
+        R"(rot_icrs_itrs(xyz: typing.Iterable[expression], time_expr: expression = heyoka.time, thresh: float = {}, eop_data: eop_data = eop_data()) -> list[expression]
+
+Rotation from ICRS to ITRS.
+
+.. versionadded:: 7.3.0
+
+This function will rotate the input Cartesian coordinates *xyz* from the axes of the
+`ICRS <https://en.wikipedia.org/wiki/International_Celestial_Reference_System_and_its_realizations>`__ frame to the axes of the
+`ITRS <https://en.wikipedia.org/wiki/International_Terrestrial_Reference_System_and_Frame>`__ frame.
+
+While the orientation of the ICRS is fixed in inertial space, the orientation of the ITRS is fixed relative to the Earth,
+and thus the rotation depends on the input time *time_expr* (which is expected to represent the number of Julian centuries elapsed
+since the epoch of J2000 in the `terrestrial time scale (TT) <https://en.wikipedia.org/wiki/Terrestrial_Time>`__).
+
+The rotation also requires in input an Earth orientation parameters dataset *eop_data* and a truncation threshold *thresh*
+for the IAU2000/2006 precession-nutation theory. Please see the documentation of :py:class:`~heyoka.eop_data` and
+:py:func:`~heyoka.model.iau2006()` for more information.
+
+:param xyz: the input Cartesian coordinates.
+:param time_expr: the input time expression.
+:param thresh: the truncation threshold for the IAU2000/2006 theory.
+:param eop_data: the EOP data to be used for the computation.
+
+:returns: the rotated Cartesian coordinates.
+
+)",
+        thresh);
 }
 
 std::string era()
@@ -1586,9 +1654,9 @@ std::string dYp()
     return detail::dXp_dYp_impl("Y");
 }
 
-std::string iau2006()
+std::string iau2006(double thresh)
 {
-    return R"(iau2006(time_expr: expression = heyoka.time, thresh: float = 1e-6) -> list[expression]
+    return fmt::format(R"(iau2006(time_expr: expression = heyoka.time, thresh: float = {}) -> list[expression]
 
 IAU2000/2006 precession-nutation theory.
 
@@ -1623,7 +1691,8 @@ will be discarded. In order to formulate the full theory without truncation, use
 
 :raises ValueError: it *thresh* is negative or non-finite.
 
-)";
+)",
+                       thresh);
 }
 
 } // namespace heyoka_py::docstrings