+++ doc/changelog.rst
@@ -15,7 +15,7 @@ New
   (`#222 <https://github.com/bluescarni/heyoka.py/pull/222>`__).
 - New models: time conversions, frame transformations, Earth orientation
   parameters (including ERA, polar motion, etc.)
-  (`#226 <https://github.com/bluescarni/heyoka.py/pull/226>`__,
+  (`#225 <https://github.com/bluescarni/heyoka.py/pull/225>`__,
   `#223 <https://github.com/bluescarni/heyoka.py/pull/223>`__,
   `#222 <https://github.com/bluescarni/heyoka.py/pull/222>`__).